@@ -13,10 +13,10 @@
     "axios": "^0.21.1",
     "bootstrap": "^5.0.2",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.0.0-beta.4",
     "react-bootstrap-icons": "^1.5.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-geocode": "^0.2.3",
     "react-lodash": "^0.1.2",
     "react-redux": "^7.2.4",
